requirements.txt
numpy>=1.24
Pillow>=9.0
torch>=1.10  # For torch.cat compatibility
